Improve debug output during app creation
